requirements.txt
openai==1.3.7
google-generativeai==0.8.3
httpx==0.24.1
python-dotenv==1.0.0
pydantic==2.5.0
beautifulsoup4==4.12.2
requests==2.31.0 